+++ app/models/user_copy.py
@@ -12,8 +12,6 @@ class User(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(40), nullable=False, unique=True)
     email = db.Column(db.String(255), nullable=False, unique=True)
-    first_name = db.Column(db.String(75), nullable=False)
-    last_name = db.Column(db.String(75), nullable=False)
     hashed_password = db.Column(db.String(255), nullable=False)
 
     @property
@@ -31,7 +29,5 @@ class User(db.Model, UserMixin):
         return {
             'id': self.id,
             'username': self.username,
-            'email': self.email,
-            "first_name": self.first_name,
-            "last_name": self.last_name
+            'email': self.email
         }
