--- host-a-hound/.flaskenv
@@ -1,2 +1,0 @@
-FLASK_APP=app
-FLASK_ENV=development
