--- host-a-hound/.gitignore
@@ -1,6 +1,0 @@
-.env
-__pycache__/
-*.py[cod]
-.venv
-.DS_Store
-.vscode/
