--- host-a-hound/Pipfile
@@ -1,29 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-click = "*"
-gunicorn = "*"
-itsdangerous = "*"
-python-dotenv = "*"
-six = "*"
-Flask = "*"
-Flask-Cors = "*"
-Flask-SQLAlchemy = "*"
-Flask-WTF = "*"
-SQLAlchemy = "*"
-Werkzeug = "*"
-WTForms = "*"
-Flask-Migrate = "*"
-Flask-Login = "*"
-alembic = "*"
-python-dateutil = "*"
-python-editor = "*"
-Mako = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.9"
