--- host-a-hound/app/models/__init__.py
@@ -1,3 +1,0 @@
-from .db import db
-from .user import User
-from .db import environment, SCHEMA
