--- host-a-hound/app/models/db.py
@@ -1,15 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-import os
-environment = os.getenv("FLASK_ENV")
-SCHEMA = os.environ.get("SCHEMA")
-
-
-db = SQLAlchemy()
-
-# helper function for adding prefix to foreign key column references in production
-def add_prefix_for_prod(attr):
-    if environment == "production":
-        return f"{SCHEMA}.{attr}"
-    else:
-        return attr
