--- host-a-hound/react-app/.env.example
@@ -1,1 +1,0 @@
-REACT_APP_BASE_URL=http://localhost:5000
