--- host-a-hound/react-app/public/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Simple React App</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
