--- host-a-hound/react-app/src/components/auth/LogoutButton.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { useDispatch } from 'react-redux';
-import { logout } from '../../store/session';
-
-const LogoutButton = () => {
-  const dispatch = useDispatch()
-  const onLogout = async (e) => {
-    await dispatch(logout());
-  };
-
-  return <button onClick={onLogout}>Logout</button>;
-};
-
-export default LogoutButton;
