--- host-a-hound/react-app/src/index.css
@@ -1,1 +1,0 @@
-/* TODO Add site wide styles */
