--- host-a-hound/react-app/src/index.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import './index.css';
-import App from './App';
-import configureStore from './store';
-
-const store = configureStore();
-
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-        <App />
-      </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
