--- host-a-hound/requirements.txt
@@ -1,24 +1,0 @@
--i https://pypi.org/simple
-alembic==1.8.1
-click==8.1.3
-flask==2.2.2
-flask-cors==3.0.10
-flask-login==0.6.2
-flask-migrate==4.0.0
-flask-sqlalchemy==3.0.2
-flask-wtf==1.0.1
-gunicorn==20.1.0
-importlib-metadata==5.0.0 ; python_version < '3.10'
-itsdangerous==2.1.2
-jinja2==3.1.2 ; python_version >= '3.7'
-mako==1.2.4
-markupsafe==2.1.1 ; python_version >= '3.7'
-python-dateutil==2.8.2
-python-dotenv==0.21.0
-python-editor==1.0.4
-setuptools==65.6.0 ; python_version >= '3.7'
-six==1.16.0
-sqlalchemy==1.4.44
-werkzeug==2.2.2
-wtforms==3.0.1
-zipp==3.10.0 ; python_version >= '3.7'
